jenkins

+++ jenkins.win.bat
@@ -11,14 +11,15 @@ set KONDA_ARGS=^
 conda\%JOB_BASE_NAME%
 set KONDA_TOKEN=C:\Users\WIN8\anaconda-token.txt
 
-%KONDA% env remove --yes --prefix env
-
-rem# if you have ssl connection problems with it
-rem# copy files "libcrypto-1_1-x64.dll" and "libssl-1_1-x64.dll"
-rem# from "Library/bin" to "DDLs"
-%KONDA% create --yes --prefix env
-
-%KONDA% install --prefix env conda-build anaconda-client
+if exist env (
+) else (
+    %KONDA% env remove --yes --prefix env
+    rem# if you have ssl connection problems with it
+    rem# copy files "libcrypto-1_1-x64.dll" and "libssl-1_1-x64.dll"
+    rem# from "Library/bin" to "DDLs"
+    %KONDA% create --yes --prefix env
+    %KONDA% install --prefix env conda-build anaconda-client
+)
 
 rem FOR /F "tokens=*" %%g IN ('env\Scripts\conda build --output %KONDA_ARGS%') do (SET KONDA_PAK=%%g)
 FOR /F "tokens=*" %%g IN ('env\condabin\conda build --output %KONDA_ARGS%') do (SET KONDA_PAK=%%g)
